consumer/requirements.txt
# Required Python libraries

psycopg2-binary==2.9.7
pika==1.3.2
pandas==1.3.5